Complete migration of Monster Registry to Go.

+++ atlas.com/morg/models/damage_summary.go
@@ -1,9 +1,9 @@
 package models
 
 type DamageSummary struct {
-	CharacterId   int
-	Monster       Monster
-	VisibleDamage int32
-	ActualDamage  int32
-	Killed        bool
+   CharacterId   int
+   Monster       Monster
+   VisibleDamage int64
+   ActualDamage  int64
+   Killed        bool
 }
